Addressed PR feedback. Extracted duplicate test classes and methods, renamed acronym for clarity.
