+++ test/Microsoft.IdentityModel.Tokens.Saml.Tests/Telemetry/SamlSignatureValidationTelemetryTests.cs
@@ -3,12 +3,10 @@
 
 using System;
 using System.Collections.Generic;
-using System.Diagnostics.Metrics;
-using System.Linq;
-using System.Security.Cryptography;
 using Microsoft.IdentityModel.Logging;
 using Microsoft.IdentityModel.Telemetry;
 using Microsoft.IdentityModel.TestUtils;
+using Microsoft.IdentityModel.TestUtils.Telemetry;
 using Xunit;
 
 namespace Microsoft.IdentityModel.Tokens.Saml.Tests.Telemetry;
@@ -49,7 +47,7 @@ public class SamlSignatureValidationTelemetryTests
         handler.ValidateTokenAsync(tokenString, validationParameters);
 
         // Assert
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -96,7 +94,7 @@ public class SamlSignatureValidationTelemetryTests
         }
 
         // Assert
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -135,7 +133,7 @@ public class SamlSignatureValidationTelemetryTests
         handler.ValidateTokenAsync(tokenString, validationParameters);
 
         // Assert
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -182,7 +180,7 @@ public class SamlSignatureValidationTelemetryTests
         }
 
         // Assert
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -221,7 +219,7 @@ public class SamlSignatureValidationTelemetryTests
         handler.ValidateTokenAsync(tokenString, validationParameters);
 
         // Assert
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -260,7 +258,7 @@ public class SamlSignatureValidationTelemetryTests
         handler.ValidateTokenAsync(tokenString, validationParameters);
 
         // Assert
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -299,7 +297,7 @@ public class SamlSignatureValidationTelemetryTests
         handler.ValidateTokenAsync(tokenString, validationParameters);
 
         // Assert
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -338,7 +336,7 @@ public class SamlSignatureValidationTelemetryTests
         handler.ValidateTokenAsync(tokenString, validationParameters);
 
         // Assert
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -347,101 +345,4 @@ public class SamlSignatureValidationTelemetryTests
                 { TelemetryConstants.KeySizeTag, 2048 }
             });
     }
-
-    private void AssertTelemetryRecorded(TestMeterListener listener, string counterName, Dictionary<string, object> expectedTags)
-    {
-        var measurements = listener.GetMeasurements(counterName);
-        Assert.NotEmpty(measurements);
-
-        var lastMeasurement = measurements.Last();
-        foreach (var expectedTag in expectedTags)
-        {
-            var actualTag = lastMeasurement.Tags.FirstOrDefault(t => t.Key == expectedTag.Key);
-            Assert.NotNull(actualTag.Value);
-            Assert.Equal(expectedTag.Value.ToString(), actualTag.Value.ToString());
-        }
-    }
-
-    /// <summary>
-    /// Test meter listener to capture telemetry measurements
-    /// </summary>
-    private class TestMeterListener : System.IDisposable
-    {
-        private readonly MeterListener _listener;
-        private readonly List<Measurement> _measurements = new List<Measurement>();
-        private readonly long _startTimestamp;
-
-        public TestMeterListener()
-        {
-            _startTimestamp = System.Diagnostics.Stopwatch.GetTimestamp();
-            _listener = new MeterListener();
-            _listener.InstrumentPublished = (instrument, listener) =>
-            {
-                if (instrument.Meter.Name == "MicrosoftIdentityModel_Meter")
-                {
-                    listener.EnableMeasurementEvents(instrument);
-                }
-            };
-
-            _listener.SetMeasurementEventCallback<long>((instrument, measurement, tags, state) =>
-            {
-                lock (_measurements)
-                {
-                    _measurements.Add(new Measurement
-                    {
-                        InstrumentName = instrument.Name,
-                        Value = measurement,
-                        Tags = tags.ToArray(),
-                        Timestamp = System.Diagnostics.Stopwatch.GetTimestamp()
-                    });
-                }
-            });
-
-            _listener.Start();
-        }
-
-        public List<Measurement> GetMeasurements(string instrumentName)
-        {
-            lock (_measurements)
-            {
-                // Only return measurements that were recorded after this listener started
-                return _measurements
-                    .Where(m => m.InstrumentName == instrumentName && m.Timestamp >= _startTimestamp)
-                    .ToList();
-            }
-        }
-
-        public void Dispose()
-        {
-            _listener?.Dispose();
-        }
-
-        public class Measurement
-        {
-            public string InstrumentName { get; set; }
-            public long Value { get; set; }
-            public KeyValuePair<string, object>[] Tags { get; set; }
-            public long Timestamp { get; set; }
-        }
-    }
-
-    /// <summary>
-    /// Creates an RSA instance with the specified key size.
-    /// On .NET Framework, uses RSACng for better support of larger key sizes (3072, 4096).
-    /// On other platforms, uses the default RSA.Create().
-    /// </summary>
-    private static RSA CreateRsa(int keySize)
-    {
-#if NET462 || NET472
-        // Use RSACng on .NET Framework for better support of 3072/4096 bit keys
-        var rsa = new System.Security.Cryptography.RSACng();
-        rsa.KeySize = keySize;
-        return rsa;
-#else
-        var rsa = RSA.Create();
-        rsa.KeySize = keySize;
-        return rsa;
-#endif
-    }
-
 }
