+++ test/Microsoft.IdentityModel.Tokens.Tests/Telemetry/SignatureValidationAndDecryptionTelemetryTests.cs
@@ -1,16 +1,17 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
 using System.Collections.Generic;
-using System.Diagnostics.Metrics;
 using System.IdentityModel.Tokens.Jwt;
-using System.Linq;
+#if !NET462 && !NET472 && !NETSTANDARD2_0
 using System.Security.Cryptography;
+#endif
 using System.Threading.Tasks;
 using Microsoft.IdentityModel.JsonWebTokens;
 using Microsoft.IdentityModel.Logging;
 using Microsoft.IdentityModel.Telemetry;
 using Microsoft.IdentityModel.TestUtils;
+using Microsoft.IdentityModel.TestUtils.Telemetry;
 using Microsoft.IdentityModel.Tokens.Experimental;
 using Xunit;
 
@@ -49,13 +50,13 @@ public class SignatureValidationAndDecryptionTelemetryTests
 
         // Assert
         Assert.True(result.IsValid);
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
-                { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
-                { TelemetryConstants.StatusTag, TelemetryConstants.SuccessValue },
-                { TelemetryConstants.AlgorithmTag, SecurityAlgorithms.RsaSha256 },
-                { TelemetryConstants.KeySizeTag, KeyingMaterial.JsonWebKeyRsa256SigningCredentials.Key.KeySize }
+                    { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
+                    { TelemetryConstants.StatusTag, TelemetryConstants.SuccessValue },
+                    { TelemetryConstants.AlgorithmTag, SecurityAlgorithms.RsaSha256 },
+                    { TelemetryConstants.KeySizeTag, KeyingMaterial.JsonWebKeyRsa256SigningCredentials.Key.KeySize }
             });
     }
 
@@ -86,7 +87,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
 
         // Assert
         Assert.False(result.IsValid);
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -124,7 +125,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
 
         // Assert
         Assert.True(result.IsValid);
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -163,7 +164,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
 
         // Assert
         Assert.False(result.IsValid);
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -203,7 +204,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
 
         // Assert
         Assert.True(result.Succeeded);
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -244,7 +245,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
 
         // Assert
         Assert.True(result.Succeeded);
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -284,7 +285,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
 
         // Assert
         Assert.False(result.Succeeded);
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -325,7 +326,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
 
         // Assert
         Assert.False(result.Succeeded);
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -362,7 +363,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
         handler.ValidateToken(token, validationParameters, out _);
 
         // Assert
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -405,7 +406,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
         }
 
         // Assert
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -443,7 +444,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
         handler.ValidateToken(token, validationParameters, out _);
 
         // Assert
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -489,7 +490,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
         }
 
         // Assert
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -526,7 +527,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
 
         // Assert
         Assert.True(result.IsValid);
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -569,8 +570,8 @@ public class SignatureValidationAndDecryptionTelemetryTests
 
         // Assert
         Assert.True(result.IsValid);
-        // For key wrap scenarios, telemetry records the wrapping key size (RSA key) to monitor STS key upgrades (2048 → 3072 → 4096).
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
+        // For key wrap scenarios, telemetry records the wrapping key size (RSA key) to monitor STS key upgrades (2048 ? 3072 ? 4096).
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -581,25 +582,6 @@ public class SignatureValidationAndDecryptionTelemetryTests
             });
     }
 
-    /// <summary>
-    /// Creates an RSA instance with the specified key size.
-    /// On .NET Framework, uses RSACng for better support of larger key sizes (3072, 4096).
-    /// On other platforms, uses the default RSA.Create().
-    /// </summary>
-    private static RSA CreateRsa(int keySize)
-    {
-#if NET462 || NET472
-        // Use RSACng on .NET Framework for better support of 3072/4096 bit keys
-        var rsa = new System.Security.Cryptography.RSACng();
-        rsa.KeySize = keySize;
-        return rsa;
-#else
-        var rsa = RSA.Create();
-        rsa.KeySize = keySize;
-        return rsa;
-#endif
-    }
-
     [Fact]
     public async Task ValidateToken_RsaKeyWrapEncryption_3072_RecordsTelemetry()
     {
@@ -608,7 +590,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
         var handler = new JsonWebTokenHandler();
 
         // Create RSA 3072-bit key for encryption
-        using var rsa = CreateRsa(3072);
+        using var rsa = KeyingMaterial.CreateRsa(3072);
         var rsaKey = new RsaSecurityKey(rsa) { KeyId = "Rsa3072Key" };
 
         var encryptingCredentials = new EncryptingCredentials(
@@ -639,7 +621,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
         // Assert
         Assert.True(result.IsValid);
         // Telemetry should track the 3072-bit RSA wrapping key size to monitor STS upgrades
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -658,7 +640,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
         var handler = new JsonWebTokenHandler();
 
         // Create RSA 4096-bit key for encryption
-        using var rsa = CreateRsa(4096);
+        using var rsa = KeyingMaterial.CreateRsa(4096);
         var rsaKey = new RsaSecurityKey(rsa) { KeyId = "Rsa4096Key" };
 
         var encryptingCredentials = new EncryptingCredentials(
@@ -689,7 +671,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
         // Assert
         Assert.True(result.IsValid);
         // Telemetry should track the 4096-bit RSA wrapping key size to monitor STS upgrades
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -734,7 +716,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
         // Assert
         Assert.True(result.IsValid);
         // RSA-OAEP is the recommended algorithm; telemetry should track wrapping key size
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -782,7 +764,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
         // Assert
         Assert.True(result.IsValid);
         // For direct key use, key size equals the content encryption key size
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -830,7 +812,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
         // Assert
         Assert.True(result.IsValid);
         // For AES key wrap, telemetry tracks the KEK (Key Encryption Key) size
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.TokenDecryptionCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -873,7 +855,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
 
         // Assert
         Assert.True(result.IsValid);
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -914,7 +896,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
 
         // Assert
         Assert.True(result.IsValid);
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -955,7 +937,7 @@ public class SignatureValidationAndDecryptionTelemetryTests
 
         // Assert
         Assert.True(result.IsValid);
-        AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
+        TelemetryAssertionHelpers.AssertTelemetryRecorded(listener, TelemetryDataRecorder.SignatureValidationCounterName,
             new Dictionary<string, object>
             {
                 { TelemetryConstants.IdentityModelVersionTag, IdentityModelTelemetryUtil.ClientVer },
@@ -965,81 +947,4 @@ public class SignatureValidationAndDecryptionTelemetryTests
             });
     }
 #endif
-
-    private static void AssertTelemetryRecorded(TestMeterListener listener, string counterName, Dictionary<string, object> expectedTags)
-    {
-        var measurements = listener.GetMeasurements(counterName);
-        Assert.NotEmpty(measurements);
-
-        var lastMeasurement = measurements.Last();
-        foreach (var expectedTag in expectedTags)
-        {
-            var actualTag = lastMeasurement.Tags.FirstOrDefault(t => t.Key == expectedTag.Key);
-            Assert.NotNull(actualTag.Value);
-            Assert.Equal(expectedTag.Value.ToString(), actualTag.Value.ToString());
-        }
-    }
-
-    /// <summary>
-    /// Test meter listener to capture telemetry measurements
-    /// </summary>
-    private class TestMeterListener : System.IDisposable
-    {
-        private readonly MeterListener _listener;
-        private readonly List<Measurement> _measurements = new List<Measurement>();
-        private readonly long _startTimestamp;
-
-        public TestMeterListener()
-        {
-            _startTimestamp = System.Diagnostics.Stopwatch.GetTimestamp();
-            _listener = new MeterListener();
-            _listener.InstrumentPublished = (instrument, listener) =>
-            {
-                if (instrument.Meter.Name == "MicrosoftIdentityModel_Meter")
-                {
-                    listener.EnableMeasurementEvents(instrument);
-                }
-            };
-
-            _listener.SetMeasurementEventCallback<long>((instrument, measurement, tags, state) =>
-            {
-                lock (_measurements)
-                {
-                    _measurements.Add(new Measurement
-                    {
-                        InstrumentName = instrument.Name,
-                        Value = measurement,
-                        Tags = tags.ToArray(),
-                        Timestamp = System.Diagnostics.Stopwatch.GetTimestamp()
-                    });
-                }
-            });
-
-            _listener.Start();
-        }
-
-        public List<Measurement> GetMeasurements(string instrumentName)
-        {
-            lock (_measurements)
-            {
-                // Only return measurements that were recorded after this listener started
-                return _measurements
-                    .Where(m => m.InstrumentName == instrumentName && m.Timestamp >= _startTimestamp)
-                    .ToList();
-            }
-        }
-
-        public void Dispose()
-        {
-            _listener?.Dispose();
-        }
-
-        public class Measurement
-        {
-            public string InstrumentName { get; set; }
-            public long Value { get; set; }
-            public KeyValuePair<string, object>[] Tags { get; set; }
-            public long Timestamp { get; set; }
-        }
-    }
 }
